Separate into field_element
