+++ src/models.rs
@@ -1,3 +1,4 @@
 pub mod field;
+pub mod field_element;
 pub mod field_object;
 pub mod game;
